07_Feature_Store/scripts/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pandas==2.0.2
numpy==1.25.2
pyyaml==6.0
sqlalchemy==2.0.19
redis==4.6.0
